Save Photo to Cloud
This change actually saves the photo to Firebase.  It also updates the
Todo Details view so that it displays the image using the downloadURL
from Firebase.

However, it does not yet save the Item object’s photos field to the
database, nor does it get deleted when an Item is deleted.  Changes
also need to be made to get it to work right when editing an item.