Redirect User to Todo List Upon Login
Previously, when a user logged in, they were taken back to the login
page.

Now, when a user logs in, they are taken to the Todo list.

Previously, when a user logged in, logged out, then logged back in, the
logout button was blank.

Now, under these circumstances, the log out button correctly shows:
Logout: <user email>